+++ front/profile.form.php
@@ -7,6 +7,8 @@ if (!defined('GLPI_ROOT')) {
 use GlpiPlugin\Ticketmigration\Menu;
 use GlpiPlugin\Ticketmigration\MigrationProfile;
 use GlpiPlugin\Ticketmigration\ProfileRight;
+use GlpiPlugin\Ticketmigration\WebUrl;
+use GlpiPlugin\Ticketmigration\SourceFile;
 
 if (!ProfileRight::canManageProfiles(UPDATE)) {
     Html::displayErrorAndDie(__('You do not have permission to perform this action.'));
@@ -27,15 +29,26 @@ $id = (int) ($_GET['id'] ?? 0);
 if ($id > 0 && (!$profile->getFromDB($id) || !$profile->canViewItem())) {
     Html::displayErrorAndDie(__('You do not have permission to view this migration profile.', 'ticketmigration'));
 }
+$activeSource = null;
+if ($id > 0 && (int) ($profile->fields['sourcefiles_id'] ?? 0) > 0) {
+    $source = new SourceFile();
+    if ($source->getFromDB((int) $profile->fields['sourcefiles_id']) && $source->canViewItem()) {
+        $activeSource = $source->fields;
+    }
+}
 
-Html::header(__('Migration profile', 'ticketmigration'), $_SERVER['PHP_SELF'], 'tools', Menu::class);
+Html::header(__('Migration profile details', 'ticketmigration'), $_SERVER['PHP_SELF'], 'tools', Menu::class);
 Glpi\Application\View\TemplateRenderer::getInstance()->display(
     '@ticketmigration/profile/form.html.twig',
     [
         'profile' => $profile->fields,
         'profile_id' => $id,
         'form_action' => MigrationProfile::getFormURL(),
-        'upload_url' => $CFG_GLPI['root_doc'] . '/plugins/ticketmigration/front/source.form.php',
+        'upload_url' => WebUrl::front('source.form.php'),
+        'sources_url' => WebUrl::front('source.php'),
+        'mapping_url' => WebUrl::front('mapping.form.php'),
+        'preview_url' => WebUrl::front('preview.php'),
+        'active_source' => $activeSource,
         'active_entity' => Session::getActiveEntity(),
     ],
 );